refactor(mister)!: drop the inherited v2 suffix from the module path
The standalone library was github.com/ZaparooProject/zaparoo-core/mister/v2
only because it sits under a repo whose module is v2. Nothing about the library
is on its second major version, and the suffix would have locked it into a
mister/v2.x.y tag series for good.

Rename it to github.com/ZaparooProject/zaparoo-core/mister so it can start at
v0, where the API is free to move while it settles. Nothing depends on it yet
and no version has been tagged, so this costs nobody anything today and would
have been breaking the moment someone pinned it.

The root require is now a v0.0.0 placeholder; the replace means it is never
fetched, and it gets a real version once the first tag lands.
